+++ noodles-sam/src/header/record/value/map/tag/other.rs
@@ -31,6 +31,12 @@ impl<S> PartialEq for Other<S> {
     }
 }
 
+impl<S> PartialEq<[u8; LENGTH]> for Other<S> {
+    fn eq(&self, other: &[u8; LENGTH]) -> bool {
+        self.0.eq(other)
+    }
+}
+
 impl<S> Eq for Other<S> {}
 
 impl<S> fmt::Display for Other<S> {